tests: convert a few test application to 64bits pointers

+++ tests/buttons/main.c
@@ -44,37 +44,37 @@
 #if defined (BTN0_PIN) || defined (BTN1_PIN) || defined (BTN2_PIN) || defined (BTN3_PIN)
 static void cb(void *arg)
 {
-    printf("Pressed BTN%d\n", (int)arg);
+    printf("Pressed BTN%" PRIiPTR "\n", (uintptr_t)arg);
 }
 #endif
 
 int main(void)
 {
-    int cnt = 0;
+    unsigned cnt = 0;
     /* get the number of available buttons and init interrupt handler */
 #ifdef BTN0_PIN
-    if (gpio_init_int(BTN0_PIN, BTN0_MODE, BTN0_INT_FLANK, cb, (void *)cnt) < 0) {
+    if (gpio_init_int(BTN0_PIN, BTN0_MODE, BTN0_INT_FLANK, cb, (void*)(uintptr_t)cnt) < 0) {
         puts("[FAILED] init BTN0!");
         return 1;
     }
     ++cnt;
 #endif
 #ifdef BTN1_PIN
-    if (gpio_init_int(BTN1_PIN, BTN1_MODE, BTN1_INT_FLANK, cb, (void *)cnt) < 0) {
+    if (gpio_init_int(BTN1_PIN, BTN1_MODE, BTN1_INT_FLANK, cb, (void*)(uintptr_t)cnt) < 0) {
         puts("[FAILED] init BTN1!");
         return 1;
     }
     ++cnt;
 #endif
 #ifdef BTN2_PIN
-    if (gpio_init_int(BTN2_PIN, BTN2_MODE, BTN2_INT_FLANK, cb, (void *)cnt) < 0) {
+    if (gpio_init_int(BTN2_PIN, BTN2_MODE, BTN2_INT_FLANK, cb, (void*)(uintptr_t)cnt) < 0) {
         puts("[FAILED] init BTN2!");
         return 1;
     }
     ++cnt;
 #endif
 #ifdef BTN3_PIN
-    if (gpio_init_int(BTN3_PIN, BTN3_MODE, BTN3_INT_FLANK, cb, (void *)cnt) < 0) {
+    if (gpio_init_int(BTN3_PIN, BTN3_MODE, BTN3_INT_FLANK, cb, (void*)(uintptr_t)cnt) < 0) {
         puts("[FAILED] init BTN3!");
         return 1;
     }
@@ -88,7 +88,7 @@ int main(void)
         puts("[FAILED] no buttons available!");
         return 2;
     }
-    printf(" -- Available buttons: %i\n\n", cnt);
+    printf(" -- Available buttons: %d\n\n", cnt);
     puts(" -- Try pressing buttons to test.\n");
     puts("[SUCCESS]");
     return 0;
